requirements.txt
python-telegram-bot
telethon
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
ruff
pre-commit 